fiz sync when not logged

+++ public/get/load/sync.php
@@ -1,10 +1,12 @@
 <?php
 
-$dir = PATH_HOME . "_cdn/sync/" . $_SESSION['userlogin']['id'];
+if(!empty($_SESSION['userlogin'])) {
+    $dir = PATH_HOME . "_cdn/sync/" . $_SESSION['userlogin']['id'];
 
-if(file_exists($dir)) {
-    foreach (\Helpers\Helper::listFolder($dir) as $item) {
-        $data['data'][str_replace(".json", "", $item)] = json_decode(file_get_contents($dir . "/" . $item), !0);
-        unlink($dir . "/" . $item);
+    if (file_exists($dir)) {
+        foreach (\Helpers\Helper::listFolder($dir) as $item) {
+            $data['data'][str_replace(".json", "", $item)] = json_decode(file_get_contents($dir . "/" . $item), !0);
+            unlink($dir . "/" . $item);
+        }
     }
 }
